requirements.txt
# # Add your project dependencies here
# # Core Dependencies
# streamlit>=1.31.0
# langchain>=0.1.9
# langchain-community>=0.0.24
# openai>=1.12.0
# python-dotenv>=1.0.0

# # Database
# sqlalchemy>=2.0.27
# psycopg2-binary>=2.9.9
# alembic>=1.13.1

# # Vector Store
# faiss-cpu>=1.12.0
# sentence-transformers>=2.3.1
# chromadb>=0.4.22

# # Data Processing
# xarray>=2024.1.0
# netCDF4>=1.6.5
# pandas>=2.2.0
# numpy>=1.26.3

# # Visualization
# plotly>=5.18.0
# folium>=0.15.1
# streamlit-folium>=0.17.0
# matplotlib>=3.8.2

# # API
# fastapi>=0.109.2
# uvicorn>=0.27.1
# pydantic>=2.6.1

# # Utilities
# requests>=2.31.0
# tqdm>=4.66.1
# pyyaml>=6.0.1

# Core Dependencies
streamlit==1.31.0
langchain==0.1.9
langchain-community==0.0.24
openai==1.12.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Store
faiss-cpu==1.12.0
sentence-transformers==2.3.1
chromadb==0.4.22

# Data Processing
xarray==2024.1.0
netCDF4==1.6.5
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.17.0
matplotlib==3.8.2

# API
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Utilities
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1
